renamed network_data_source package
