+++ app/src/main/kotlin/com/igor_shaula/api_polling/ThisApp.kt
@@ -9,7 +9,7 @@ import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.preferencesDataStore
 import com.igor_shaula.api_polling.data_layer.AbstractVehiclesRepository
-import com.igor_shaula.api_polling.data_layer.network.NetworkRepositoryImpl
+import com.igor_shaula.api_polling.data_layer.network_data_source.NetworkRepositoryImpl
 import com.igor_shaula.api_polling.data_layer.stub_data_source.StubRepositoryImpl
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
