+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/Extensions.kt
@@ -1,7 +1,7 @@
 package com.igor_shaula.api_polling.data_layer
 
-import com.igor_shaula.api_polling.data_layer.network.VehicleDetailsModel
-import com.igor_shaula.api_polling.data_layer.network.VehicleModel
+import com.igor_shaula.api_polling.data_layer.network_data_source.VehicleDetailsModel
+import com.igor_shaula.api_polling.data_layer.network_data_source.VehicleModel
 
 fun List<Pair<String, VehicleRecord>>.toVehicleRecordList(): List<VehicleRecord> {
     val result = mutableListOf<VehicleRecord>()
