+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/network_data_source/Models.kt
@@ -1,4 +1,4 @@
-package com.igor_shaula.api_polling.data_layer.network
+package com.igor_shaula.api_polling.data_layer.network_data_source
 
 import com.google.gson.annotations.SerializedName
