+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/network_data_source/NetworkRepositoryImpl.kt
@@ -1,9 +1,9 @@
-package com.igor_shaula.api_polling.data_layer.network
+package com.igor_shaula.api_polling.data_layer.network_data_source
 
 import com.igor_shaula.api_polling.data_layer.AbstractVehiclesRepository
 import com.igor_shaula.api_polling.data_layer.VehicleDetailsRecord
 import com.igor_shaula.api_polling.data_layer.VehicleRecord
-import com.igor_shaula.api_polling.data_layer.network.retrofit.VehicleRetrofitNetworkServiceImpl
+import com.igor_shaula.api_polling.data_layer.network_data_source.retrofit.VehicleRetrofitNetworkServiceImpl
 import com.igor_shaula.api_polling.data_layer.toVehicleItemRecords
 
 class NetworkRepositoryImpl : AbstractVehiclesRepository() {
